Restructure consts and interfaces
